requirements.txt
pandas~=2.3.0
numpy
scikit-learn~=1.7.0
implicit~=0.7.2
scipy~=1.15.3
faiss-cpu~=1.11.0
sentence-transformers~=4.1.0
lightgbm~=4.6.0
torch~=2.7.1
datasets
torch-geometric~=2.6.1
fastapi~=0.116.1
pydantic~=2.11.7
uvicorn
pika==1.3.2
psycopg
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1
redis